Implement PoC of Git credential helper
The git-credential helper is implemented using async and the Tokio
runtime as that is what the Octrocrab crate uses.

The command is configured using the following environment variables:
* GITHUB_APP_ID: the ID of the Github app used for authentication
* GITHUB_APP_PRIVATE_KEY: the path to the private key for the app
* GITHUB_URL (optional): the URL of the Github API (default:
  https://api.github.com)

Tokens are currently not cached and a repository clone will lead to
multiple calls of the credential helper which will alway request and
return a new installation auth token.

This means that we could theoretically run into rate limits.